Added wait method first daft, and moved the failing tests,
The difference in Guzzle promise implementations mainly lay in the construction.

According to https://github.com/promises-aplus/constructor-spec/issues/18 it's not valid.
The constructor starts the fate of the promise. Which has to been delayed, under Guzzle.

The wait method is necessary in certain callable functions situations.
The constructor will fail it's execution when trying to access member method that's null.
It will happen when passing an promise object not fully created, itself.

Normally an promise is attached to an external running event loop, no need to start the process.
The wait function/method both starts and stops it, internally.